refactor(security): split scanner.rs into statements, expressions, checks, analysis
